bot/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-decouple==3.8
requests==2.31.0
urllib3==2.0.6
